+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - New view for selecting additional context
+  - Optionally files, mentioned in the .gitignore are automatically excluded from the context selection
 
 ## [0.1.4]